sched/fair: Allow nohz idle to "trump smp nice" too
The condition altered here is intended to avoid under-utilizing the
system due to asymmetry between the loads on groups. It was intended
to solve the case where one high-priority task is given an entire
sched_group to itself while several low-priority tasks are forced to
be co-scheduled on another. However it also solves the case where the
groups have asymmetrical capacity (such as in big.LITTLE systems).

The exception is currently only present for CPU_NEWLY_IDLE. This
means that to solve the underutilization the idle CPU must wake up
and go back to sleep. Therefore allow the exception to take place for
CPU_IDLE too.

Change-Id: I807ba5cba0ef1b8bbec02cbcd4755fd32af10135
Signed-off-by: Brendan Jackman <brendan.jackman@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>